requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pyjwt==2.8.0
pydantic==1.10.13
python-multipart==0.0.9
